+++ src/pygeoprocessing/geoprocessing.py
@@ -112,16 +112,16 @@ def raster_calculator(
 
     Parameters:
         base_raster_path_band_const_list (list): a list containing either
-            (str, int) tuples, numbers, or `numpy.ndarray`s of up to two
-            dimensions. A `(str, int)` tuple refers to a raster path band
-            index pair to use as an input. Numbers and `numpy.ndarray`s
-            are constants.
-            The rasters in this list must have the same raster size and
-            `numpy.ndarray`s must be numpy broadcastable to those rasters.
-            If only constants are input, the numpy arrays must be
-            broadcastable to each other and the final raster size will be the
-            final broadcast array shape. It is possible to make a 1x1 raster
-            with only scalar constant inputs.
+            (str, int) tuples, `numpy.ndarray`s of up to two
+            dimensions, or an (object, 'raw') tuple.  A `(str, int)`
+            tuple refers to a raster path band index pair to use as an input.
+            The `numpy.ndarray`s must be broadcastable to each other AND the
+            size of the raster inputs. Values passed by  `(object, 'raw')`
+            tuples pass `object` directly into the `local_op`. All rasters
+            must have the same raster size. If only arrays are input, numpy
+            arrays must be broadcastable to each other and the final raster
+            size will be the final broadcast array shape. A value error is
+            raised if only "raw" inputs are passed.
         local_op (function) a function that must take in as many parameters as
             there are elements in `base_raster_path_band_const_list`. The
             parameters in `local_op` will map 1-to-1 in order with the values
@@ -180,13 +180,16 @@ def raster_calculator(
     else:
         for value in base_raster_path_band_const_list:
             if (not _is_raster_path_band_formatted(value) and
-                    not isinstance(value, (numbers.Number, numpy.ndarray))):
+                not isinstance(value, numpy.ndarray) and
+                not (isinstance(value, tuple) and len(value) == 2 and
+                     value[1] == 'raw')):
                 bad_raster_path_list = True
                 break
     if bad_raster_path_list:
         raise ValueError(
-            "Expected a list of path / integer band tuples, numbers, or "
-            "ndarrays for `base_raster_path_band_const_list`, instead got: "
+            "Expected a list of path / integer band tuples, "
+            "ndarrays, or (value, 'raw') pairs for "
+            "`base_raster_path_band_const_list`, instead got: "
             "%s" % pprint.pformat(base_raster_path_band_const_list))
 
     # check that any rasters exist on disk and have enough bands
@@ -194,7 +197,7 @@ def raster_calculator(
     gdal.PushErrorHandler('CPLQuietErrorHandler')
     base_raster_path_band_list = [
         path_band for path_band in base_raster_path_band_const_list
-        if isinstance(path_band, tuple)]
+        if _is_raster_path_band_formatted(path_band)]
     for value in base_raster_path_band_list:
         if gdal.OpenEx(value[0], gdal.OF_RASTER) is None:
             not_found_paths.append(value[0])
@@ -227,7 +230,7 @@ def raster_calculator(
     raster_info_list = [
         get_raster_info(path_band[0])
         for path_band in base_raster_path_band_const_list
-        if isinstance(path_band, tuple)]
+        if _is_raster_path_band_formatted(path_band)]
     geospatial_info_set = set()
     for raster_info in raster_info_list:
         geospatial_info_set.add(raster_info['raster_size'])
@@ -288,26 +291,34 @@ def raster_calculator(
                 "Raster size %s cannot be broadcast to numpy shape %s" % (
                     raster_shape, numpy_broadcast_size))
 
-    # create a "canonical" argument list that's bands or 2d numpy arrays only
+    # create a "canonical" argument list that's bands, 2d numpy arrays, or
+    # raw values only
     base_canonical_arg_list = []
     base_raster_list = []
     base_band_list = []
     for value in base_raster_path_band_const_list:
         # the input has been tested and value is either a raster/path band
-        # tuple, 1d ndarray, 2d ndarray, or a scalar
-        if isinstance(value, tuple):
+        # tuple, 1d ndarray, 2d ndarray, or (value, 'raw') tuple.
+        if _is_raster_path_band_formatted(value):
             # it's a raster/path band, keep track of open raster and band
             # for later so we can __swig_destroy__ them.
             base_raster_list.append(gdal.OpenEx(value[0], gdal.OF_RASTER))
             base_band_list.append(
                 base_raster_list[-1].GetRasterBand(value[1]))
             base_canonical_arg_list.append(base_band_list[-1])
-        elif isinstance(value, numpy.ndarray) and value.ndim == 1:
-            # easier to process as a 2d array for writing to band
-            base_canonical_arg_list.append(value.reshape((1, value.shape[0])))
+        elif isinstance(value, numpy.ndarray):
+            if value.ndim == 1:
+                # easier to process as a 2d array for writing to band
+                base_canonical_arg_list.append(
+                    value.reshape((1, value.shape[0])))
+            else:  # dimensions are two because we checked earlier.
+                base_canonical_arg_list.append(value)
+        elif isinstance(value, tuple):
+            base_canonical_arg_list.append(value[0])
         else:
-            # otherwise it's a 2d array or scalar, pass as is
-            base_canonical_arg_list.append(value)
+            raise ValueError(
+                "An unexpected `value` occurred. This should never happen. "
+                "Value: %r" % value)
 
     # create target raster
     if raster_info_list:
@@ -321,8 +332,11 @@ def raster_calculator(
         else:
             n_rows, n_cols = numpy_broadcast_size
     else:
-        # otherwise raster is a 1x1 because of only scalar input
-        n_rows, n_cols = 1, 1
+        raise ValueError(
+            "Only (object, 'raw') values have been passed. Raster "
+            "calculator requires at least a raster or numpy array as a "
+            "parameter. This is the input list: %s" % pprint.pformat(
+                base_raster_path_band_const_list))
 
     # create target raster
     gtiff_driver = gdal.GetDriverByName('GTiff')
@@ -430,11 +444,11 @@ def raster_calculator(
             pixels_processed += blocksize[0] * blocksize[1]
             last_time = _invoke_timed_callback(
                 last_time, lambda: LOGGER.info(
-                    '%.2f%% complete',
+                    '%.1f%% complete',
                     float(pixels_processed) / n_pixels * 100.0),
                 _LOGGING_PERIOD)
 
-        LOGGER.info('100.00%% complete')
+        LOGGER.info('100.0%% complete')
 
         if calc_raster_stats:
             LOGGER.info("signaling stats worker to terminate")
@@ -1205,7 +1219,8 @@ def zonal_statistics(
             for aggregate_id in numpy.unique(valid_aggregate_id):
                 aggregate_mask = valid_aggregate_id == aggregate_id
                 masked_clipped_block = valid_clipped[aggregate_mask]
-                clipped_nodata_mask = (masked_clipped_block == raster_nodata)
+                clipped_nodata_mask = numpy.isclose(
+                    masked_clipped_block, raster_nodata)
                 if aggregate_id not in aggregate_stats:
                     aggregate_stats[aggregate_id] = {
                         'min': None,
@@ -1267,6 +1282,10 @@ def get_vector_info(vector_path, layer_index=0):
         layer_index (int): index of underlying layer to analyze.  Defaults to
             0.
 
+    Raises:
+        ValueError if `vector_path` does not exist on disk or cannot be opened
+        as a gdal.OF_VECTOR.
+
     Returns:
         raster_properties (dictionary): a dictionary with the following
             properties stored under relevant keys.
@@ -1277,7 +1296,12 @@ def get_vector_info(vector_path, layer_index=0):
                 box in projected coordinates as [minx, miny, maxx, maxy].
 
     """
-    vector = gdal.OpenEx(vector_path)
+    if not os.path.exists(vector_path):
+        raise ValueError("%s does not exist." % vector_path)
+    vector = gdal.OpenEx(vector_path, gdal.OF_VECTOR)
+    if not vector:
+        raise ValueError(
+            "Could not open %s as a gdal.OF_VECTOR" % vector_path)
     vector_properties = {}
     layer = vector.GetLayer(iLayer=layer_index)
     # projection is same for all layers, so just use the first one
@@ -1295,6 +1319,10 @@ def get_raster_info(raster_path):
 
     Parameters:
        raster_path (String): a path to a GDAL raster.
+
+    Raises:
+        ValueError if `raster_path` is not a file or cannot be opened as a
+        gdal.OF_RASTER.
 
     Returns:
         raster_properties (dictionary): a dictionary with the properties
@@ -1322,8 +1350,13 @@ def get_raster_info(raster_path):
                 efficient reading.
 
     """
+    if not os.path.exists(raster_path):
+        raise ValueError("%s does not exist." % raster_path)
+    raster = gdal.OpenEx(raster_path, gdal.OF_RASTER)
+    if not raster:
+        raise ValueError(
+            "Could not open %s as a gdal.OF_RASTER" % raster_path)
     raster_properties = {}
-    raster = gdal.OpenEx(raster_path)
     raster_properties['projection'] = raster.GetProjection()
     geo_transform = raster.GetGeoTransform()
     raster_properties['geotransform'] = geo_transform
@@ -1546,7 +1579,7 @@ def warp_raster(
         target_raster_path (string): the location of the resized and
             resampled raster.
         resample_method (string): the resampling technique, one of
-            "nearest|bilinear|cubic|cubic_spline|lanczos|average|mode|max"
+            "near|bilinear|cubic|cubicspline|lanczos|average|mode|max"
             "min|med|q1|q3"
         target_bb (list): if None, target bounding box is the same as the
             source bounding box.  Otherwise it's a list of float describing